+++ .github/workflows/check.yml
@@ -9,7 +9,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
 
-      - name: Run elastisearch container
+      - name: Run opensearch container
         run: docker compose up -d opensearch
 
       - name: Testing
@@ -24,4 +24,4 @@ jobs:
           echo "---" > ~/.gem/credentials
           echo ":github: Bearer ${{ secrets.PACKAGE_TOKEN }}" >> ~/.gem/credentials
           chmod 600 ~/.gem/credentials
-          find ./build/gems -name 'embulk-output-elasticsearch-*.gem' -maxdepth 1 -print0 | xargs -0 gem push --key github --host 'https://rubygems.pkg.github.com/calorie'
+          find ./build/gems -name 'embulk-output-opensearch-*.gem' -maxdepth 1 -print0 | xargs -0 gem push --key github --host 'https://rubygems.pkg.github.com/calorie'
