+++ build.gradle
@@ -16,7 +16,7 @@ repositories {
 
 group = "org.embulk"
 version = "0.6.0-SNAPSHOT"
-description = "Elasticsearch output plugin is an Embulk plugin that loads records to Elasticsearch read by any input plugins."
+description = "OpenSearch output plugin is an Embulk plugin that loads records to OpenSearch read by any input plugins."
 
 sourceCompatibility = JavaVersion.VERSION_11
 targetCompatibility = JavaVersion.VERSION_11
@@ -75,9 +75,9 @@ dependencies {
 }
 
 embulkPlugin {
-    mainClass = "org.embulk.output.elasticsearch.OpenSearchOutputPlugin"
+    mainClass = "org.embulk.output.opensearch.OpenSearchOutputPlugin"
     category = "output"
-    type = "elasticsearch"
+    type = "opensearch"
 }
 
 javadoc {
@@ -205,8 +205,8 @@ signing {
 gem {
     authors = [ "Muga Nishizawa" ]
     email = [ "muga.nishizawa@gmail.com" ]
-    summary = "Elasticsearch output plugin for Embulk"
-    homepage = "https://github.com/embulk/embulk-output-elasticsearch"
+    summary = "OpenSearch output plugin for Embulk"
+    homepage = "https://github.com/calorie/embulk-output-elasticsearch"
     licenses = [ "Apache-2.0" ]
 
     from rootProject.file("LICENSE")
