+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = "embulk-output-elasticsearch"
+rootProject.name = "embulk-output-opensearch"
