+++ src/main/java/org/embulk/output/opensearch/OpenSearchClientCreator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.embulk.output.elasticsearch;
+package org.embulk.output.opensearch;
 
 import org.opensearch.client.opensearch.OpenSearchClient;
