+++ src/main/java/org/embulk/output/opensearch/jackson/JacksonServiceRecord.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.embulk.output.elasticsearch.jackson;
+package org.embulk.output.opensearch.jackson;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
