+++ src/test/java/org/embulk/output/opensearch/IndexData.java
@@ -1,4 +1,4 @@
-package org.embulk.output.elasticsearch;
+package org.embulk.output.opensearch;
 
 public class IndexData
 {
